added logik to add an item to the shopping cart with increasing the quantity by adding same item multiple times

+++ spring-boot/src/main/java/com/dobri/springboot/items/ItemsRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 
 
-public interface ItemsRepository extends JpaRepository <Items, Integer> {
+public interface ItemsRepository extends JpaRepository <Items, Long> {
 }
